Component로 분할
